ConBox 기능 확장 + 데모/문서 갱신
컨트롤(Source/ConBox):
- IME 조합을 줄 중간 삽입(뒤 글자 밀어쓰기)으로 변경, 조합 중 커서는 속 빈 2px 테두리 사각형으로 표시
- 한글 모드 2칸 블록 커서가 덮는 글자를 클리핑해 다시 그려 가림 방지
- 방향키 위/아래로 화면 줄 간 이동(끈적 열 유지, 줄 끝 보정, 편집 영역 내)
- OnPaint 더블 버퍼링으로 빠른 입력 시 깜빡임 제거
- 더블클릭 단어 선택(영문/한글/특수문자 부류 기준) + 클립보드 복사
- set_margin(top,left,bottom,right) 안쪽 여백 API(기본 10, CSS 단축 해석)
- 폰트 크기 인자 int -> float(소수점 pt)
- set_kfont_match_efont 옵션(기본 켜짐): 한글을 영문과 같은 픽셀 높이로 맞추고,
  한글은 자연 폭을 유지한 채 영문 칸 폭을 좁혀(영문 장평) 한글 자간을 줄임

데모(Project):
- 확인/취소 버튼 제거, 타이틀바 최소화/최대화 버튼 추가, ConBox가 사방 20px로 채움
- System DPI Aware 적용(InitInstance에서 SetProcessDPIAware)

문서:
- 요구사양서/설계명세서/개발노트에 위 변경 반영

+++ Source/Accessories.h
@@ -17,4 +17,4 @@
 //             숫자+W = Weight (예 "700W" 면 Weight 를 700 으로)
 //           숫자는 바로 뒤에 오는 속성의 값(정도)으로 쓰인다.
 // dpi_y   : 세로 방향 DPI (포인트를 픽셀 높이로 환산할 때 사용, 보통 96)
-LOGFONTW ParseFontOpts(const char* name, int size_pt, const char* opts, int dpi_y);
+LOGFONTW ParseFontOpts(const char* name, float size_pt, const char* opts, int dpi_y);
